# machineLearning/static/swagger.json
{"openapi": "3.0.0", "info": {"title": "Solar Production Prediction API", "description": "API pour pr\u00e9dire la production d'\u00e9nergie solaire \u00e0 partir de pr\u00e9visions m\u00e9t\u00e9orologiques", "version": "1.0.0"}, "servers": [{"url": "/"}], "tags": [{"name": "Prediction", "description": "Endpoints pour les pr\u00e9dictions"}, {"name": "Model", "description": "Informations sur le mod\u00e8le"}], "paths": {"/predict": {"post": {"tags": ["Prediction"], "summary": "Pr\u00e9dire la production d'\u00e9nergie solaire", "description": "Prend des donn\u00e9es m\u00e9t\u00e9orologiques pr\u00e9dictives et renvoie la production d'\u00e9nergie solaire pr\u00e9dite", "requestBody": {"description": "Donn\u00e9es m\u00e9t\u00e9o \u00e0 utiliser pour la pr\u00e9diction", "required": true, "content": {"multipart/form-data": {"schema": {"type": "object", "properties": {"file": {"type": "string", "format": "binary", "description": "Fichier CSV contenant les donn\u00e9es m\u00e9t\u00e9orologiques"}}}}, "application/json": {"schema": {"type": "array", "items": {"type": "object", "properties": {"datetime": {"type": "string", "format": "date-time", "description": "Date et heure de la pr\u00e9diction"}, "PRED_GLOB_ctrl": {"type": "number", "description": "Pr\u00e9diction de rayonnement solaire global"}, "PRED_T_2M_ctrl": {"type": "number", "description": "Pr\u00e9diction de temp\u00e9rature \u00e0 2m"}, "PRED_RELHUM_2M_ctrl": {"type": "number", "description": "Pr\u00e9diction d'humidit\u00e9 relative \u00e0 2m"}, "PRED_TOT_PREC_ctrl": {"type": "number", "description": "Pr\u00e9diction de pr\u00e9cipitations totales"}}, "required": ["datetime", "PRED_GLOB_ctrl", "PRED_T_2M_ctrl", "PRED_RELHUM_2M_ctrl", "PRED_TOT_PREC_ctrl"]}}}}}, "responses": {"200": {"description": "Pr\u00e9dictions r\u00e9ussies", "content": {"application/json": {"schema": {"type": "object", "properties": {"predictions": {"type": "array", "items": {"type": "object", "properties": {"datetime": {"type": "string", "format": "date-time"}, "predicted_pv_output": {"type": "number"}}}}, "model_info": {"type": "object", "properties": {"training_date": {"type": "string"}, "model_type": {"type": "string"}, "metrics": {"type": "object"}}}}}}}}, "400": {"description": "Donn\u00e9es invalides", "content": {"application/json": {"schema": {"type": "object", "properties": {"error": {"type": "string"}}}}}}, "500": {"description": "Erreur serveur", "content": {"application/json": {"schema": {"type": "object", "properties": {"error": {"type": "string"}}}}}}}}}, "/model-info": {"get": {"tags": ["Model"], "summary": "Informations sur le mod\u00e8le", "description": "Renvoie des informations sur le mod\u00e8le charg\u00e9", "responses": {"200": {"description": "Informations sur le mod\u00e8le", "content": {"application/json": {"schema": {"type": "object", "properties": {"features": {"type": "array", "items": {"type": "string"}}, "training_date": {"type": "string"}, "model_type": {"type": "string"}, "metrics": {"type": "object"}}}}}}, "500": {"description": "Erreur serveur", "content": {"application/json": {"schema": {"type": "object", "properties": {"error": {"type": "string"}}}}}}}}}}}